media player: fix rapid seeking breaking internal video time sync (?)
It appears that this issue is resolved by not stacking video seeks on
top of each other, and instead waiting for each seek to finish before
starting another. Either way, seek speed seems much faster with this
"lock" implemented, and I notice no difference in perceived seek
accuracy. So hopefully a win-win.