post: remove gfycat and imgur embeds
Gfycat is shutting down on 2023-09-01.

Imgur has indicated they want to start deleting anonymous uploads and
NSFW content (the two most important use cases for Redditors). Unclear
if this is actually happening yet, but furthermore their embed code only
works half the time anyways, so it's a good chance to excise it.