frontpage: save 0.5s load time by pulling 1X subs instead of 2X
Previously we randomized the 49 frontpage subs out of a pool of the 98
most popular subs on each load - the reason being to have more variety
in the subs shown. However the load time cost of this is just too much
over simply pulling and showing the top 49 exactly. Plus, it's arguable
that always showing the "top" of the most-popular list is a better user
experience anyway.